Random generation updates
Now the user is asked the number of buildings and animals to spawn randomly. Special care is taken to generate buildings and animals properly, but currently, buildings can intersect. Also, players and animals can no longer go vertically indefinitely. Perhaps someday it might be interesting to make it possible to do so horizontally and vertically.