Added death and destruction
The player can kill deer by walking onto them, and the background color for that tile is changed to red. Deer can currently kill each other. I fixed that, but it makes the game way more interesting if they can, so I reverted the changes and kept The Great Deer War perpetuated... for now...