Fixed creatures moving more than once in a turn
This long-standing bug was caused by incorrect updating of the m boolean, which allows a tile to move randomly. Now, each tile's "mobility" is calculated before drawing each frame, and the m boolean for each tile is untouched otherwise, except for in kill(), making sure that the killer can not move again.