Removed deprecated rmove() capability
rmove() once had a way of increasing the number of random moves per turn, but it can not be enforced due to a code change intended to fix multiple moves per turn. I would've fixed this, but I'm not intending to use it anyway, so it will be re-implemented when the time comes.