Added construction and corpses
Tiles such as walls, doors, floors, and grass can be created by pressing m, followed by the first letter of the tile you want to make, and then a directional key. Period is accepted as meaning 'here'. For example, typing mwy 'm'akes a 'w'all northwest of the player. mgj 'm'akes 'g'rass south of the player, etc.

Also, when things die, they now leave a corpse as the background of the tile they were killed on. Corpses appear as red versions of the characters of the creature they represented.